requirements.txt
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-RESTful
pandas
WTForms
pyJWT
